Neurons/requirements.txt
fastapi
uvicorn[standard]
pillow
openai
python-dotenv
PyMuPDF
python-multipart